+++ Assignment-4/FinanceTracker/Model/Income.cs
@@ -1,5 +1,7 @@
+using FinanceTracker.Controller;
+using FinanceTracker.Utilities;
 
-﻿namespace FinanceTracker
+namespace FinanceTracker.Model
 {
     internal class Income
     {
